# Change Management
It’s not enough to simply let the user change the values of an object’s properties directly. Because we need to sync changes to the server, we have to track changes, and because we want to let the user jump around without worrying about state, we need to persist those changes locally until they are merged.

## GHManagedObject (Change)

This category exposes an unordered to-many relationship of LEChange objects, as well as methods for setting and getting change values.

Users of GHManagedObjects can then access and mutate their change values, as distinct from their original values, which will continue to reflect the truth from the server.

## LEChange

Changes are embodied by the LEChange model object, which exposes a local property name, a remote key name, and the change, stored as a string value for simplicity.

Change objects are not handled directly, but are exposed through methods on GHManagedObject.

## NSDate (GitHub)

Upgrade this convenience category to a two-way transformer between GitHub-formatted date strings and NSDate objects.