# Jump from talk list to editor
A new experience that feels natural, and makes the extra accessory not just an indicator of comments, but a convenient shortcut for the less likely use case of wanting to edit the issue directly and immediately.

## LETalkListController

1. When an issue has comments, selecting that issue goes into the talk view controller.
  * From here selecting any talk goes into the work view controller (preview / editor)
2. When an issue has no comments, selecting that issue jumps into the work view controller.
3. Issue with comments show the detail disclosure accessory, which jumps as if it had no comments.

## Storyboard

The talk list now has its own segue as well as the same segue as the talk view, which allows it to jump over the talk view, as nicely displayed by the storyboard editor.

Refs #21