In GitHub, replies are comments, and comments are, at least in this use case—attached to issues.
In the UI, issues and comments both contain some common conceptual sub-object called a talk, which has a body. (Seeing the problem as Composition versus Inheritance, Refs #31)

In the code, issues and comments are both kinds of talks, their common API unified behind the LETalk protocol.

In any case, the user may be replying to the top-level issue or one of its comments, but they both result in the same thing: a new comment being added to the issue.

Therefore, we have to disambiguate between whether the current talk is the issue, or the current talk is contained by the issue. For this we can introspect the actual class of the talk, and use the issue back pointer if necessary.

## Next steps

1. Create a comment and attach it to the issue.
2. Ensure this is done in a way that is compatible with the coming sync architecture.
3. Actually sync these changes.

Refs #30