# Pre-structuring
The narrative of the experience and its UI are starting to come together, as the tension between GItHub and Twitter begin to tear the model apart. Refs #31

There is a list of conversations, displayed by topics, containing talks by users.

The structure of the project should reflect this.

Unfortunately, changing the names and locations of things causes tears in the history, so we’ve postponing that, while trying on the new structure.

Once the experience gels, and everything is committed, the filesystem can be changed to reflect the project structure, and it can all be folded into the master at one time.